Update
# wireFrame Mode Switch 기능 추가
# InputHandler Excute 대상 Camera에서 Actor로 변경
# Camrea 방향 회전 Quternion에서 일반 회전으로 변경
# Grid 추가
# png, jpg에 대한 텍스쳐 생성 추가

+++ GraphicsStudy/Renderer/StaticMesh.h
@@ -31,7 +31,7 @@ namespace Core {
 			m_objectConstantData->Model = m_objectConstantData->Model.Transpose();
 			std::vector<ObjectConstantData> constantData = { *m_objectConstantData	};
 			Renderer::Utility::CreateUploadBuffer(constantData, m_objectConstantBuffer, device);
-			indexCount = meshData.m_indices.size();
+			indexCount = (UINT)meshData.m_indices.size();
 
 			CD3DX12_RANGE range(0, 0);
 			ThrowIfFailed(m_objectConstantBuffer->Map(0, &range, reinterpret_cast<void**>(&m_pCbvDataBegin)));
